this should be an int (member variable is)

+++ java/code/src/com/suse/manager/model/products/Product.java
@@ -99,7 +99,7 @@ public class Product implements Selectable, Comparable<Product> {
          * Get the product synchronization progress.
          * @return the progress
          */
-        public float getSyncProgress() {
+        public int getSyncProgress() {
             return syncProgress;
         }
